refactor: update tag feature
