+++ src/app/mod.rs
@@ -1,3 +1,2 @@
 pub mod article;
 pub mod comment;
-pub mod tag;
