--- src/app/tag/mod.rs
@@ -1,3 +1,0 @@
-pub mod api;
-pub mod model;
-pub mod response;
