+++ src/appv2/features/favorite/presenters.rs
@@ -1,10 +1,8 @@
 use actix_web::HttpResponse;
 
 pub use crate::app::article::response::SingleArticleResponse;
-use crate::{
-    app::{article::model::Article, tag::model::Tag},
-    appv2::features::profile::entities::Profile,
-};
+use crate::appv2::features::tag::entities::Tag;
+use crate::{app::article::model::Article, appv2::features::profile::entities::Profile};
 
 use super::entities::FavoriteInfo;
