+++ src/appv2/features/favorite/repositories.rs
@@ -1,8 +1,8 @@
 use super::entities::FavoriteInfo;
 use super::services;
 use crate::app::article::model::Article;
-use crate::app::tag::model::Tag;
 use crate::appv2::features::profile::entities::Profile;
+use crate::appv2::features::tag::entities::Tag;
 use crate::appv2::features::user::entities::User;
 use crate::error::AppError;
 use crate::utils::db::DbPool;
