+++ src/appv2/features/mod.rs
@@ -2,4 +2,5 @@ pub mod favorite;
 pub mod follow;
 pub mod healthcheck;
 pub mod profile;
+pub mod tag;
 pub mod user;
