+++ src/appv2/features/tag/mod.rs
@@ -1,0 +1,3 @@
+pub mod controllers;
+pub mod entities;
+pub mod presenters;
